requirements.txt
kivy>=2.2.0
kivymd>=1.1.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
pillow>=10.0.0
